moved more location stuff
