+++ lib/features/driver/data/models/update_driver_location_dto.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:malawi_ride_share_app/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/models/driver_status.dart';
-import 'package:malawi_ride_share_app/features/driver/data/models/location/location_dto.dart';
+import 'package:malawi_ride_share_app/features/location/data/models/location_dto.dart';
 
 part 'update_driver_location_dto.freezed.dart';
 part 'update_driver_location_dto.g.dart';
