+++ lib/features/driver/domain/repository/driver_location_tracking_repository.dart
@@ -1,4 +1,4 @@
-import 'package:malawi_ride_share_app/features/driver/data/models/location/location_dto.dart';
+import 'package:malawi_ride_share_app/features/location/data/models/location_dto.dart';
 import 'package:malawi_ride_share_app/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/models/driver_status.dart';
 
 abstract class DriverLocationTrackingRepository {
