+++ lib/features/location/domain/entities/location.dart
